add index file and spec; rename sensor to node
