+++ nomad.json
@@ -1,7 +1,3 @@
-// "QmYcbm7GPAd3S7nSSrsxuyxef9KQ7fR3SkTmLcWX9BdN8y"   // resolves (is an id)
-// "Qmf8Ps1gfrkDRXjF2vsBwEbThczvPepSzXUA3yh64aSVD6"   // does not resolve
-// "Qmc6cSnvbGUfiJaUmu5tX4AzX1HUqdrokLvDMJk5gihQ76" is a content hash
-
 {
   "subscriptions": [
     "QmZa2gmiDLVLkcUgz5oL36KTK3pL4Jf2VNkmhJWjjkNWwk"
