+++ src/index.js
@@ -1,0 +1,3 @@
+const Node = require('./node')
+
+module.exports = Node
