+++ src/subscribe.js
@@ -1,3 +1,9 @@
+// TODO: remove keys
+// "QmYcbm7GPAd3S7nSSrsxuyxef9KQ7fR3SkTmLcWX9BdN8y"   // resolves (is an id)
+// "Qmf8Ps1gfrkDRXjF2vsBwEbThczvPepSzXUA3yh64aSVD6"   // does not resolve
+// "Qmc6cSnvbGUfiJaUmu5tX4AzX1HUqdrokLvDMJk5gihQ76" is a content hash
+
+
 const R = require('ramda')
 
 const log = require('./utils/log')
